refactor(editor): restructure TipTap editor utilities and improve code organization
- Move tiptap-utils.ts to modular lib/editor/utils/ structure
- Add @tiptap/core and @tiptap/extension-emoji dependencies
- Clean up CSS classes and improve code formatting
- Update 35 files with consistent improvements
- Remove 394 lines of legacy utility code
- Add proper exports in editor index.ts

This refactoring improves maintainability and follows better code organization patterns.
