+++ lib/editor/components/ui/heading-button/heading-button.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 
 // --- Lib ---
-import { parseShortcutKeys } from "@/lib/tiptap-utils"
+import { parseShortcutKeys } from "@/lib/editor/utils"
 
 // --- Tiptap UI ---
 import type {
