+++ lib/editor/components/ui/text-align-button/text-align-button.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 
 // --- Lib ---
-import { parseShortcutKeys } from "@/lib/tiptap-utils"
+import { parseShortcutKeys } from "@/lib/editor/utils"
 
 // --- Hooks ---
 import { useTiptapEditor } from "@/lib/editor/hooks/use-tiptap-editor"
