+++ lib/editor/index.ts
@@ -3,3 +3,4 @@ export * from './extensions';
 export * from './providers';
 export * from './hooks';
 export * from './contents';
+export * from './utils';
