+++ package.json
@@ -64,6 +64,8 @@
 		"@tanstack/react-query-devtools": "^5.81.5",
 		"@tanstack/react-table": "^8.21.3",
 		"@tanstack/react-virtual": "^3.13.12",
+		"@tiptap/core": "^3.6.1",
+		"@tiptap/extension-emoji": "^3.6.1",
 		"@tiptap/extension-highlight": "^3.4.4",
 		"@tiptap/extension-horizontal-rule": "^3.4.4",
 		"@tiptap/extension-image": "^3.4.4",
